backend/requirements-minimal.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google AI (Gemini) - direct SDK without langchain
google-generativeai==0.8.0

# E2B Code Interpreter for secure code execution
e2b-code-interpreter==1.5.2

# Basic utilities
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic>=2.7.0
requests==2.31.0 